+++ CHANGELOG.md
@@ -1,5 +1,179 @@
 # Start From 1.6.7 version
 
+## 1.6.10
+
+- Increase performance
+
+- Remove `IOLevelLog.ALL`
+
+- Added `events.EventEmitter` with 7 events:
+
+  - API interface `IOKeyEvents` to get 7 events:
+
+    - `levelChange`
+    - `logging`
+    - `fatalLogging`
+    - `settingChange`
+    - `childCreated`
+    - `willLog`
+    - `loggerNameChange`
+
+- Bring back `onInit()`
+
+  ```ts
+  const logger = new Logger(
+    {
+      format: "pretty",
+      levelLog: Number.POSITIVE_INFINITY,
+      useColor: true,
+    },
+    (Logger) => {
+      Logger.info("Logger is ready");
+    }
+  );
+  ```
+
+- Changed `stacktrace.ts` to `error-stack-parser` package
+
+- `loggerProps`, `childProps` and `getBindingOpt()` now is read-only
+
+  ```ts
+  const logger = new Logger({
+    format: "pretty",
+    levelLog: Number.POSITIVE_INFINITY,
+    useColor: true,
+  });
+
+  const childLogger = logger.child({ a: "b" }, { isChild: true });
+
+  childLogger.getBindingOpt().a; /* Nothing happend */
+  childLogger.getBindingOpt().a = "c"; /* This line will throw Error */
+
+  childLogger.loggerProps.isChild; /* Nothing happend */
+  childLogger.loggerProps.isChild = false; /* This line will throw Error */
+  ```
+
+- Added `isChild`, `parentName`, `parentOldProps` and new `child()` method
+
+  ```ts
+  export class Logger<P extends {}, OP extends {} = {}> extends LoggerUtils<P> {
+    private _isChild: boolean = false;
+    private _parentName: string = "";
+    public parentOldProps: Readonly<OP> = {} as Readonly<OP>;
+    public get isChild(): boolean {
+      return this._isChild;
+    }
+    public get parentName(): string {
+      return this._parentName;
+    }
+    /* <Some logic code . . .> */
+
+    public child<T extends {}, LP extends {} = {}>(
+      bindingOpt?: T,
+      loggerOpt?: LP,
+      childSetting: IOLoggerInterface = this.listSetting()
+    ): ChildLogger<P, T, LP> {
+      const childLogger: Logger<P & T, P> = new Logger<P & T, P>({
+        ...this.listSetting(),
+        ...childSetting,
+      });
+
+      const loggerProps = Object.freeze({
+        loggerProps: Object.freeze(loggerOpt),
+      } as IOChildLoggerProperty<LP>);
+
+      let bindingLoggerProps = Object.assign(childLogger, loggerProps);
+
+      bindingLoggerProps.isChild = true;
+      bindingLoggerProps.childProps = {
+        ...this.childProps,
+        ...(bindingOpt ?? ({} as T)),
+      };
+      bindingLoggerProps.parentName = this.loggerName;
+      bindingLoggerProps.parentOldProps = this.getBindingOpt();
+
+      this.emit(
+        "childCreated",
+        this,
+        bindingLoggerProps,
+        childSetting,
+        childSetting.instanceName || this.listSetting().instanceName || "",
+        Object.freeze(bindingOpt!),
+        loggerProps
+      );
+      return bindingLoggerProps;
+    }
+  }
+  ```
+
+- Added `getBindingOpt()` to get `childProps`
+
+  ```ts
+  const logger = new Logger({
+    format: "pretty",
+    levelLog: Number.POSITIVE_INFINITY,
+    useColor: true,
+  });
+
+  const childLogger = logger.child({ a: "b" }, { isChild: true });
+
+  console.log("Binding:", childLogger.getBindingOpt().a);
+
+  /*----------------Output----------------*/
+  /*
+  Binding: b
+  */
+  ```
+
+- Edited `IOErrorStack`
+
+  ```ts
+  interface IOErrorStack {
+    filePath: string;
+    fullFilePath: string | undefined;
+    lineNumber: number | undefined;
+    lineColumm: number | undefined;
+    functionName: string | undefined;
+    methodName: string | undefined;
+    isConstructor: boolean | undefined;
+  }
+  ```
+
+- Added `useExpressLogger()` for `express`
+
+  ```ts
+  import express from "express";
+  import Logger from "@khoa1468/logger";
+  import { useExpressLogger } from "@khoa1468/logger";
+
+  const app = express();
+
+  const logger = new Logger();
+
+  app.use(
+    useExpressLogger(
+      logger /* Your logger here */,
+      logger.listSetting() /* Your setting of the logger */
+    )
+  );
+  ```
+
+- `levelLog` is now a `number` not an `Array`
+
+  ```ts
+  protected levelLog: IOLevelLog = IOLevelLog.NONE;
+  ```
+
+- Added custom level log for `prefix()` logging method
+
+  ```ts
+  logger.prefix({
+    prefix: "Prefix" /* <Your Prefix> */,
+    color: "red" /* <Your Color> */,
+    levelLog: 100 /* <Your Level> */,
+  });
+  ```
+
 ## 1.6.9
 
 - Small fixed, increase perfomance